+++ src/doitintl/docops/gloss/cli/terms.py
@@ -21,6 +21,7 @@
 # SOFTWARE.
 
 
+# TODO: Colorize this output
 """Find terms that may be candidates for inclusion in a glossary
 
 Usage:
@@ -28,7 +29,8 @@ Usage:
   {cmd_name} [options] <DIR>
   {cmd_name} [options] (-h | --help)
   {cmd_name} [options] (--version)
-  {cmd_name} [options] (--show-formats)
+  {cmd_name} [options] (--show-output-types)
+  {cmd_name} [options] (--show-table-formats)
   {cmd_name} [options] (--print-cache)
   {cmd_name} [options] (--delete-cache)
 
@@ -70,11 +72,15 @@ Basic options:
   -c --config-dir=DIR      Search `DIR` for `.gloss.conf` files instead of the
                            target directory
 
-  -l, --row-limit=NUM      Limit results to `NUM` rows [default: {row_limit}]
+  -l, --term-limit=NUM     Limit results to `NUM` terms [default: {term_limit}]
+
+  -o, --output-type=TYPE   Use output TYPE [default: table]
+
+  --show-output-types      Print a list of supported output types and exit
 
   -t, --table-format=TYPE  Use table format `TYPE` [default: {table_format}]
 
-  --show-formats           Print a list of supported table formats and exit
+  --show-table-formats     Print a list of supported table formats and exit
 
   --print-cache            Print the location of the cache directory and exit
 
@@ -104,7 +110,7 @@ cmd_path = pathlib.PurePath(sys.argv[0])
 cmd_name = cmd_path.name
 format_dict = {
     "cmd_name": cmd_name,
-    "row_limit": docops.row_limit,
+    "term_limit": docops.term_limit,
     "table_format": docops.table_format,
 }
 # Not needed for now, but may need again the future
@@ -117,6 +123,9 @@ __doc__ = __doc__.format(**format_dict)
 
 class TermsCommand:
 
+    OUTPUT_TYPES = ["none", "table"]
+    TABLE_FORMATS = tabulate.tabulate_formats
+
     # Allowed options
     help = None
     version = None
@@ -124,9 +133,11 @@ class TermsCommand:
     quiet = None
     disable_ansi = None
     config_dir = None
-    row_limit = None
+    output_type = None
+    show_output_types = None
+    term_limit = None
     table_format = None
-    show_formats = None
+    show_table_formats = None
     print_cache = None
     delete_cache = None
     show_warnings = None
@@ -151,11 +162,14 @@ class TermsCommand:
                 self.set_class_attr(attr_name, value)
                 continue
             attr_name = key.lower()
-            self.set_class_attr(attr_name, value)
+            self.set_class_attr(attr_name, attr_name)
         self.set_pkg_options()
         self._config = Configuration(self.dir, self.config_dir)
         self._cache = Cache()
         self._validate_docstring()
+
+    def _print(self, *args):
+        self._config.printer.print(*args)
 
     def _validate_docstring(self):
         for line_num, line in enumerate(__doc__.splitlines()):
@@ -181,7 +195,17 @@ class TermsCommand:
         docops.quiet = self.quiet
         docops.disable_ansi = self.disable_ansi
         docops.table_format = self.table_format
-        docops.row_limit = self.row_limit
+        docops.term_limit = self.term_limit
+
+    def validate_options(self):
+        if self.output_type not in self.OUTPUT_TYPES:
+            raise docops.ConfigurationError(
+                f"Not a supported type: {self.output_type}"
+            )
+        if self.table_format not in self.TABLE_FORMATS:
+            raise docops.ConfigurationError(
+                f"Not a supported type: {self.output_type}"
+            )
 
     def run(self):
         if self.help:
@@ -192,12 +216,15 @@ class TermsCommand:
             return self.do_delete_cache()
         if self.print_cache:
             return self.do_print_cache()
-        if self.show_formats:
-            return self.do_show_formats()
+        if self.show_output_types:
+            return self.do_show_output_types()
+        if self.show_table_formats:
+            return self.do_show_table_formats()
+        self.validate_options()
         self.do_main()
 
     def do_help(self):
-        print(__doc__.strip())
+        self._print(__doc__.strip())
         return 0
 
     def do_delete_cache(self):
@@ -209,20 +236,32 @@ class TermsCommand:
             return 1
         return 0
 
-    def do_show_formats(self):
-        printer = Printer()
-        msg = "Supported table formats:\n\n"
-        for format in tabulate.tabulate_formats:
-            msg += f" - {format}\n"
-        printer.print(msg.strip())
-        return 0
+    # TODO: Convert most methods of this class to private methods
+    def _do_show_options(self, plural_noun, noun_list):
+        msg = f"Supported {plural_noun}:\n\n"
+        for noun in noun_list:
+            if noun:
+                msg += f"  - <str>{noun}</str>\n"
+        self._print(msg.strip())
+
+    def do_show_output_types(self):
+        self._do_show_options("output types", self.OUTPUT_TYPES)
+
+    def do_show_table_formats(self):
+        self._do_show_options("table formats", self.TABLE_FORMATS)
 
     def do_main(self):
         self._config.load()
         self._analyzer = TermAnalyzer(self._config)
         self._analyzer.run()
-        self._analyzer.print_ranks()
-        return 0
+        if docops.verbose:
+            print("Output: {self.output_type}")
+        if self.output_type == "table":
+            self._analyzer.print_table()
+            return
+        if self.output_type == "none":
+            return
+        raise docops.ProgramError("I don't know what to do now...")
 
 
 def run():