+++ src/doitintl/docops/gloss/print.py
@@ -27,18 +27,64 @@ import pastel
 
 from doitintl import docops
 
+# For the full list of options, see:
+# https://github.com/sdispater/pastel/blob/master/pastel/style.py
 
-class Printer:
 
-    MAX_LINE_LENGTH=79
+class Printer(pastel.Pastel):
+
+    MAX_LINE_LENGTH = 79
+
+    # TODO: Make use of these throughout the code
+    STYLES = {
+        "success": {
+            "fg": "green",
+            "options": None,
+        },
+        "error": {
+            "fg": "red",
+            "options": ["bold"],
+        },
+        "warning": {
+            "fg": "yellow",
+        },
+        # Value types
+        "type": {
+            "options": ["bold"],
+        },
+        # Numeric values
+        "num": {
+            "fg": "magenta",
+        },
+        # String values
+        "str": {
+            "fg": "cyan",
+        },
+        # Object values
+        "obj": {
+            "options": ["dark"],
+        },
+        "emphasis": {
+            "options": ["italic"],
+        },
+        "strong": {
+            "options": ["bold"],
+        },
+        "special": {
+            "options": ["underline"],
+        },
+    }
 
     def __init__(self):
-        if docops.disable_ansi:
-            pastel.with_colors(False)
-        else:
-            pastel.with_colors(True)
+        super().__init__()
+        self.with_colors(not docops.disable_ansi)
+        self._configure_styles()
+
+    def _configure_styles(self):
+        for name, kwargs in self.STYLES.items():
+            self.add_style(name, **kwargs)
 
     def print(self, msg, file=sys.stdout):
-        msg = pastel.colorize(str(msg))
+        msg = self.colorize(str(msg))
         msg = f"{msg}\n"
         file.write(msg)